skin-type-api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
numpy<2.0
python-multipart


